Implement naive threading
Added a worker thread for loading entries. While it improves the
performance when loading very large databases, its effectiveness is
limited by a bottle neck in the json library: The library we use is not
threaded and has to finish parsing the entire file before the
load_database function can proceed.